# api-gw/api/goclient/v1/api_exercise.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api_exercise.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "ExerciseService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/class/{classroomID}/exercise": {
      "get": {
        "operationId": "ExerciseService_GetAllExercisesOfClassroom",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetAllExercisesOfClassroomResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "classroomID",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "titleSearch",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortColumn",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "isDesc",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "tags": [
          "ExerciseService"
        ]
      }
    },
    "/api/class/{classroomID}/exercise/stage/{categoryID}": {
      "get": {
        "operationId": "ExerciseService_GetAllExercisesInReportingStage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetAllExercisesInReportingStageResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "classroomID",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "categoryID",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "ExerciseService"
        ]
      }
    },
    "/api/exercise": {
      "get": {
        "operationId": "ExerciseService_GetExercises",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetExercisesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "titleSearch",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortColumn",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "isDesc",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "tags": [
          "ExerciseService"
        ]
      },
      "post": {
        "operationId": "ExerciseService_CreateExercise",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateExerciseResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateExerciseRequest"
            }
          }
        ],
        "tags": [
          "ExerciseService"
        ]
      }
    },
    "/api/exercise/{id}": {
      "get": {
        "operationId": "ExerciseService_GetExercise",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetExerciseResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "ExerciseService"
        ]
      },
      "delete": {
        "operationId": "ExerciseService_DeleteExercise",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteExerciseResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "ExerciseService"
        ]
      },
      "put": {
        "operationId": "ExerciseService_UpdateExercise",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateExerciseResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "exercise": {
                  "$ref": "#/definitions/v1ExerciseInput"
                }
              },
              "title": "====== PUT ======"
            }
          }
        ],
        "tags": [
          "ExerciseService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "typeDateTime": {
      "type": "object",
      "properties": {
        "year": {
          "type": "integer",
          "format": "int32",
          "description": "Optional. Year of date. Must be from 1 to 9999, or 0 if specifying a\ndatetime without a year."
        },
        "month": {
          "type": "integer",
          "format": "int32",
          "description": "Required. Month of year. Must be from 1 to 12."
        },
        "day": {
          "type": "integer",
          "format": "int32",
          "description": "Required. Day of month. Must be from 1 to 31 and valid for the year and\nmonth."
        },
        "hours": {
          "type": "integer",
          "format": "int32",
          "description": "Required. Hours of day in 24 hour format. Should be from 0 to 23. An API\nmay choose to allow the value \"24:00:00\" for scenarios like business\nclosing time."
        },
        "minutes": {
          "type": "integer",
          "format": "int32",
          "description": "Required. Minutes of hour of day. Must be from 0 to 59."
        },
        "seconds": {
          "type": "integer",
          "format": "int32",
          "description": "Required. Seconds of minutes of the time. Must normally be from 0 to 59. An\nAPI may allow the value 60 if it allows leap-seconds."
        },
        "nanos": {
          "type": "integer",
          "format": "int32",
          "description": "Required. Fractions of seconds in nanoseconds. Must be from 0 to\n999,999,999."
        },
        "utcOffset": {
          "type": "string",
          "description": "UTC offset. Must be whole seconds, between -18 hours and +18 hours.\nFor example, a UTC offset of -4:00 would be represented as\n{ seconds: -14400 }."
        },
        "timeZone": {
          "$ref": "#/definitions/typeTimeZone",
          "description": "Time zone."
        }
      },
      "description": "Represents civil time (or occasionally physical time).\n\nThis type can represent a civil time in one of a few possible ways:\n\n * When utc_offset is set and time_zone is unset: a civil time on a calendar\n   day with a particular offset from UTC.\n * When time_zone is set and utc_offset is unset: a civil time on a calendar\n   day in a particular time zone.\n * When neither time_zone nor utc_offset is set: a civil time on a calendar\n   day in local time.\n\nThe date is relative to the Proleptic Gregorian Calendar.\n\nIf year is 0, the DateTime is considered not to have a specific year. month\nand day must have valid, non-zero values.\n\nThis type may also be used to represent a physical time if all the date and\ntime fields are set and either case of the `time_offset` oneof is set.\nConsider using `Timestamp` message for physical time instead. If your use\ncase also would like to store the user's timezone, that can be done in\nanother field.\n\nThis type is more flexible than some applications may want. Make sure to\ndocument and validate your application's limitations."
    },
    "typeTimeZone": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "IANA Time Zone Database time zone, e.g. \"America/New_York\"."
        },
        "version": {
          "type": "string",
          "description": "Optional. IANA Time Zone Database version number, e.g. \"2019a\"."
        }
      },
      "description": "Represents a time zone from the\n[IANA Time Zone Database](https://www.iana.org/time-zones)."
    },
    "v1AttachmentExerciseInput": {
      "type": "object",
      "properties": {
        "fileURL": {
          "type": "string"
        },
        "authorID": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "exerciseID": {
          "type": "string",
          "format": "int64"
        },
        "size": {
          "type": "string",
          "format": "int64"
        },
        "type": {
          "type": "string"
        },
        "thumbnail": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "v1AttachmentExerciseResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "fileURL": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "author": {
          "$ref": "#/definitions/v1AuthorExerciseResponse"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "size": {
          "type": "string",
          "format": "int64"
        },
        "mimeType": {
          "type": "string"
        },
        "thumbnail": {
          "type": "string"
        },
        "fileName": {
          "type": "string"
        }
      }
    },
    "v1AuthorExerciseResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "class": {
          "type": "string"
        },
        "major": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "photoSrc": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "v1CommentExerciseResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "user": {
          "$ref": "#/definitions/v1AuthorExerciseResponse"
        },
        "exerciseID": {
          "type": "string",
          "format": "int64"
        },
        "content": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1CommonExerciseResponse": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "string",
          "format": "int64"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "v1CreateExerciseRequest": {
      "type": "object",
      "properties": {
        "exercise": {
          "$ref": "#/definitions/v1ExerciseInput"
        }
      }
    },
    "v1CreateExerciseResponse": {
      "type": "object",
      "properties": {
        "response": {
          "$ref": "#/definitions/v1CommonExerciseResponse"
        }
      }
    },
    "v1DeleteExerciseResponse": {
      "type": "object",
      "properties": {
        "response": {
          "$ref": "#/definitions/v1CommonExerciseResponse"
        }
      }
    },
    "v1ExerciseInput": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "classroomID": {
          "type": "string",
          "format": "int64"
        },
        "deadline": {
          "$ref": "#/definitions/typeDateTime"
        },
        "categoryID": {
          "type": "string",
          "format": "int64"
        },
        "authorID": {
          "type": "string"
        },
        "attachments": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1AttachmentExerciseInput"
          }
        }
      },
      "title": "====== exercise ======"
    },
    "v1ExerciseResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "classroomID": {
          "type": "string",
          "format": "int64"
        },
        "deadline": {
          "$ref": "#/definitions/typeDateTime"
        },
        "category": {
          "$ref": "#/definitions/v1ReportingStageExerciseResponse"
        },
        "author": {
          "$ref": "#/definitions/v1AuthorExerciseResponse"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "attachments": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1AttachmentExerciseResponse"
          }
        }
      }
    },
    "v1GetAllExercisesInReportingStageResponse": {
      "type": "object",
      "properties": {
        "response": {
          "$ref": "#/definitions/v1CommonExerciseResponse"
        },
        "totalCount": {
          "type": "string",
          "format": "int64"
        },
        "exercises": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1ExerciseResponse"
          }
        }
      }
    },
    "v1GetAllExercisesOfClassroomResponse": {
      "type": "object",
      "properties": {
        "totalCount": {
          "type": "string",
          "format": "int64"
        },
        "exercises": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1ExerciseResponse"
          }
        }
      }
    },
    "v1GetExerciseResponse": {
      "type": "object",
      "properties": {
        "exercise": {
          "$ref": "#/definitions/v1ExerciseResponse"
        },
        "comments": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1CommentExerciseResponse"
          }
        }
      }
    },
    "v1GetExercisesResponse": {
      "type": "object",
      "properties": {
        "totalCount": {
          "type": "string",
          "format": "int64"
        },
        "exercises": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1ExerciseResponse"
          }
        }
      }
    },
    "v1ReportingStageExerciseResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "label": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "v1UpdateExerciseResponse": {
      "type": "object",
      "properties": {
        "response": {
          "$ref": "#/definitions/v1CommonExerciseResponse"
        }
      }
    }
  }
}